november-2021-leetcoding-challenge/src/main/java/org/redquark/leetcoding/challenge/Problem11_MinimumValueToGetPositiveStepByStepSum.java
package org.redquark.leetcoding.challenge;

/**
 * @author Anirudh Sharma
 * <p>
 * Given an array of integers nums, you start with an initial positive value startValue.
 * <p>
 * In each iteration, you calculate the step by step sum of startValue plus elements in nums (from left to right).
 * <p>
 * Return the minimum positive value of startValue such that the step by step sum is never less than 1.
 * <p>
 * Constraints:
 * <p>
 * 1 <= nums.length <= 100
 * -100 <= nums[i] <= 100
 */
public class Problem11_MinimumValueToGetPositiveStepByStepSum {

    public int minStartValue(int[] nums) {
        // Special case
        if (nums == null || nums.length == 0) {
            return 0;
        }
        // Minimum prefix sum
        int minPrefixSum = 0;
        // Cumulative sum
        int sum = 0;
        // Loop through all the elements
        for (int num : nums) {
            sum += num;
            minPrefixSum = Math.min(minPrefixSum, sum);
        }
        return 1 - minPrefixSum;
    }
}
